requirements.txt
python-telegram-bot==13.7
urllib3==1.26.18
upstox-python-sdk
requests
python-dotenv
apscheduler
pytz
